fix(routing): IT4 レビューの高優先度 6 件を直す
5 視点のレビューで、**全緑のまま通っていた欠陥**が 4 件見つかった。いずれも
検査を壊しても赤にならないことを実測で確認したうえで直している。

**① すでに出港した便が候補に出る**（利用者代表・テスター）
航海スケジュールの一覧は既定で「本日以降」に絞っているのに、経路探索には
出発の下限が無く、過去の便が候補になっていた。古い便ほど日数計算上は早く着く
ため上位を占める。押さえられない船を前提にした経路が 1 位に出れば、経路設計者は
以後この一覧の順位を信用しない。Clock を注入し、一覧と同じ扱いに揃えた。

**② 同じ航海が「直行」と「1 回積み替え」の 2 候補になる**（プログラマー）
途中に寄港する 1 本の航海で、降りてまた同じ船に乗る経路を作っていた。出発も
到着も船も同じ 2 行が並び、片方だけが 25 万円高く「積み替え 1 回」と出る。
存在しない選択肢と存在しない価格差を見せていた。

**③ 積み替えの最低 6 時間が探索側で無防備**（テスター）
TransitPath.of にも同じ検査があるが、探索が先に弾くためその経路に到達しない。
探索側のガードを外しても全緑だった。境界（5 時間 59 分・ちょうど 6 時間）の
テストを探索側にも置いた。

**④ 積み替え上限の 2 つのガードが互いを隠す**（テスター）
探索側の打ち切りと条件側の判定のどちらを外しても緑だった。条件側を単独で
固定するテスト（経路を直接組み立てて探索を通さない）を足した。

**⑤ 積み替えの上限に上限が無い**（プログラマー・アーキテクト）
外から任意の値を渡せた。探索は深さに対して指数的に広がるため、1 回の
問い合わせでサービスを止められる。絶対の上限（3）を設けた。あわせて
「3 回以上は候補にしない」というコメントが ADR-018 に存在しない決定を
引用していた点も、既定値と絶対上限を分けて解消した。

**⑥ メタテストの必須規則に validationAfterAuthorizationRule が無い**（アーキテクト）
ADR-016 は「適用漏れはメタテストが検出する」と書いていたが、名簿に載って
いなかった。8 番目のサービスを足しても緑のままだった。

画面側も直した。サーバが区別した理由をそのまま見せる／期限が空なら探しに
行かず入力を促す／期限を予約から変えたら元の期限と「荷主の合意が要ります」を
残して戻せるようにする／順位が到着の早さでないことを画面に書く／待ち行列へ
戻る導線／「イテレーション」を業務の言葉に。

モックも本物に合わせた（同じ船に乗り直さない・出港済みを除く・400 の条件・
期限の端を 1 ミリ秒までそろえる）。画面が並べ替えていないことを判別できる
よう、フィクスチャの推奨順と自然順をわざと食い違わせた。

+++ apps/backend/routingms/src/main/java/com/example/routingms/domain/model/RouteRecommendation.java
@@ -62,7 +62,7 @@ public final class RouteRecommendation {
         BigDecimal legs = PER_LEG.multiply(BigDecimal.valueOf(path.edges().size()));
         BigDecimal days = PER_DAY.multiply(BigDecimal.valueOf(path.transitDays()));
         // 出発地と目的地に、経由港を足した数だけ港を使う
-        BigDecimal ports = PER_PORT.multiply(BigDecimal.valueOf(path.transitPorts().size() + 2));
+        BigDecimal ports = PER_PORT.multiply(BigDecimal.valueOf((long) path.transitPorts().size() + 2));
         return legs.add(days).add(ports);
     }
 }
